# Optimize equals