We do a lot of equality checks in our application to avoid `Xor` (and the associated allocations) if two bitmaps are equal.


Use slightly faster looping constructs for a ~20% speed up in the
normal case. Also, compare container identity to greatly speed up
comparison of cloned bitmaps.

```
benchmark                   old ns/op     new ns/op     delta
BenchmarkEqualsClone-4      148228        17037         -88.51%
BenchmarkEqualsSparse-4     166294        133839        -19.52%
```